Create ubnsystemtroubleshooting.sh
This script is a simple enough one that will help troubleshoot performance related problems on Ubuntu servers.  It can easily me edited for any Linux flavor. 

1. Clone or copy contents
2. chmod +x ubnsystemtroubleshooting.sh
3. /ubnsystemtroubleshooting.sh -h

Running the "all" flag will run everything but it will not output to your terminal.  Instead it saves a timestamped file in /tmp for analysis.

Commands will be added to grow this.